Preview files for th Django 4th hw

+++ board/views.py
@@ -16,18 +16,21 @@ Classes:
 This module makes use of Django's built-in authentication views and form handling and integrates
 models and forms from the `board` application, such as `Ad`, `Comment`, `Profile`, and `User`.
 """
-
-from django.shortcuts import render, get_object_or_404, redirect
+from django.http import HttpResponse, HttpRequest, request, Http404
+from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.contrib.auth.decorators import login_required
-from django.contrib.auth.views import LoginView
-from django.contrib.auth import login
+from django.contrib.auth.views import LoginView, LogoutView
+from django.contrib.auth import login, logout, update_session_auth_hash
 from django.contrib import messages
+from django.urls import reverse_lazy
 from django.utils import timezone
 from django.db.models import Count
 from datetime import timedelta
 
+from my_site import settings
+from my_site.settings import MEDIA_ROOT, MEDIA_URL
 from .models import Ad, User, Category, Comment, Profile
-from .forms import CommentForm, RegistrationForm
+from .forms import CommentForm, RegistrationForm, UserProfileForm, PasswordChangeForm
 
 
 def register_view(request) -> render:
@@ -43,13 +46,17 @@ def register_view(request) -> render:
     Returns:
         HttpResponse: The rendered registration page or a redirect to the login page if the form is valid.
     """
+    if request.user.is_authenticated:
+        return redirect('board:user_profile', user_id=request.user.id)
     if request.method == 'POST':
         form = RegistrationForm(request.POST, request.FILES)
         if form.is_valid():
             user = form.save()
             messages.success(request, 'Ваш акаунт успішно створено!')
-            login(request, user)
-            return redirect('board:login')
+            login(request, user, backend='django.contrib.auth.backends.ModelBackend')
+            return redirect('board:user_profile', user_id=user.id)
+        else:
+            return render(request, 'registration/register.html', {'form': form})
     else:
         form = RegistrationForm()
     return render(request, 'registration/register.html', {'form': form})
@@ -96,26 +103,32 @@ def ad_detail(request, ad_id: int) -> render:
 
 
 @login_required
-def user_profile(request, username: str) -> render:
+def user_profile(request, user_id: int) -> render:
     """
-    Displays the profile of a user.
+    Відображає профіль користувача.
 
-    Fetches the user's profile and displays it. Requires the user to be logged in.
+    Отримує профіль користувача і відображає його. Потрібно, щоб користувач був увійшов.
 
-    Args:
-        request (HttpRequest): The HTTP request object.
-        username (str): The username of the user whose profile to display.
+    Аргументи:
+        request (HttpRequest): HTTP запит.
+        user_id (int): Ідентифікатор користувача, чий профіль потрібно відобразити.
 
-    Returns:
-        HttpResponse: The rendered user profile page.
+    Повертає:
+        HttpResponse: Рендерить сторінку профілю користувача.
     """
-    user = get_object_or_404(User, username=username)
+    user = get_object_or_404(User, id=user_id)
+    if request.user != user:
+        messages.error(request, "Ви не маєте доступу до цього профілю.")
+        return redirect('board:ad_list')
+
     profile = get_object_or_404(Profile, user=user)
-    return render(request, 'board/profile.html', {'user': user, 'profile': profile})
+
+    return render(request, 'board/profile.html',
+                  {'user': user, 'profile': profile})
 
 
 @login_required
-def edit_profile(request):
+def edit_profile_view(request, user_id: int):
     """
     View for editing the user profile.
 
@@ -125,14 +138,18 @@ def edit_profile(request):
     Returns:
         HttpResponse: Renders the profile edit page with a form.
     """
-    user_profile = request.user.profile
+    user = get_object_or_404(User, id=user_id)
+    user_profile = get_object_or_404(Profile, user=user)
+
+    if request.user != user:
+        return redirect('board:user_profile', user_id=request.user.id)
 
     if request.method == "POST":
         form = UserProfileForm(request.POST, request.FILES, instance=user_profile)
         if form.is_valid():
             form.save()
             messages.success(request, "Профіль успішно оновлено!")
-            return redirect("board:profile")
+            return redirect('board:user_profile', user_id=user.id)
         else:
             messages.error(request, "Будь ласка, виправте помилки у формі.")
     else:
@@ -147,13 +164,14 @@ def change_password_view(request):
         form = PasswordChangeForm(user=request.user, data=request.POST)
         if form.is_valid():
             user = form.save()
-            update_session_auth_hash(request, user)
-            messages.success(request, "Password changed successfully.")
-            return redirect('board: profile')
-        messages.error(request, "Failed to change password. Please check the errors.")
+            update_session_auth_hash(request, user)  # Оновлює сесію для користувача після зміни пароля
+            messages.success(request, "Пароль успішно змінено.")
+            return redirect('board:user_profile', user_id=request.user.id)  # Редірект на профіль користувача
+        messages.error(request, "Не вдалося змінити пароль. Будь ласка, перевірте помилки.")
     else:
         form = PasswordChangeForm(user=request.user)
-    return render(request, 'change_password.html', {'form': form})
+
+    return render(request, 'board/change_password.html', {'form': form})
 
 
 class CustomLoginView(LoginView):
@@ -164,6 +182,16 @@ class CustomLoginView(LoginView):
         get_success_url: Returns the URL to redirect the user to after a successful login.
     """
 
+    def dispatch(self, request, *args, **kwargs):
+        """
+        Check if the user is already authenticated. If so, redirect them to their profile page.
+        """
+        if request.user.is_authenticated:
+            # Redirect to the user's profile page if already logged in
+            return redirect(reverse('board:user_profile', kwargs={'user_id': request.user.id}))
+
+        return super().dispatch(request, *args, **kwargs)
+
     def get_success_url(self) -> str:
         """
         Determines the URL to redirect to after successful login.
@@ -171,8 +199,28 @@ class CustomLoginView(LoginView):
         Returns:
             str: The URL to the user's profile page.
         """
-        print(f"Redirecting to profile for user {self.request.user.username}")
-        return f'/board/profile/{self.request.user.username}/'
+        return f'/board/profile/{self.request.user.id}/'
+
+
+class CustomLogoutView(LogoutView):
+    """
+    Custom logout view that redirects the user to the main page after logout.
+
+    Methods:
+        get_logout_url: Returns the URL to redirect the user after logout.
+    """
+
+    def get(self, request: HttpRequest, *args: tuple, **kwargs: dict) -> HttpResponse:
+        """
+        Override the default GET method to add a success message and redirect
+        the user after logging out.
+
+        Returns:
+            str: The URL to the main page.
+        """
+        messages.success(request, "Ви успішно вийшли з системи.")
+        next_page = reverse_lazy('board:ad_list')  # Using reverse_lazy to get the URL dynamically
+        return redirect(next_page)
 
 
 def ad_statistics(request) -> render:
@@ -205,9 +253,11 @@ def ad_statistics(request) -> render:
 
 
 @login_required
-def delete_account_view(request):
+def delete_account_view(request, user_id):
+    if request.user.id != user_id:
+        raise Http404("Ви не маєте прав на видалення цього акаунта.")
     if request.method == "POST":
         request.user.delete()
         logout(request)
         return redirect("board:ad_list")
-    return render(request, "board/delete_account.html")
+    return render(request, "board/delete_account.html", {'user_id': user_id})
